feat: создание упражнения

+++ src/main/resources/db/migration/common/V23011901__exercises.sql
@@ -69,6 +69,6 @@ create table images
     id         BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     name       VARCHAR,
     media_type VARCHAR NOT NULL,
-    size       INT,
+    size       BIGINT,
     data       BYTEA
 );
